Reorganize ETJump UI stuff
Not a fully exhaustive reorganization, but it's a start.

* Moved ETJump UI initalization/shutdown to 'etj_main.cpp'
* Added 'etj_local.h' for global ETJump UI objects/methods
* Added 'etj_utilities.cpp/h' for UI and moved bunch of random utility functions there
